Restructuring all scripts to do w/ experiment properties.
