+++ public/scripts/experiment-properties/balancing/latin_square_generator.js
