+++ public/scripts/experiment-properties/data/constants/jnd_data.js
@@ -1,23 +1,4 @@
-export { get_data,
-         get_data_subset };
-
-// =========================================================
-// DEFINING SUBCONDITION REPEATS
-
-const SUBCONDITION_REPEATS = {
-    foundational : {
-        base : 0
-    },
-    design : {
-        distractor_rainbow: 0
-    },
-    design_multi : {
-        distractor_multi: 4
-    }
-}
-
-// =========================================================
-// JND DATA
+export {JND_BASE, JND_CONDITIONS};
 
 const JND_BASE = {
 
@@ -118,22 +99,6 @@ const JND_BASE = {
     {distribution_type: "gaussian", base_correlation: 0.9, error: 0.0001, max_step_size: 0.01, converge_from_above: true, initial_difference: 0.1, num_points: 100, mean: 0.5, SD: 0.2, num_SD: 2.5, point_color: '#dbc667', axis_color: 'BLACK', text_color: 'BLACK', feedback_background_color: 'WHITE', background_color: 'WHITE', point_size: 3}
     
     ]
-}
-
-const JND_RADIUS = {
-
-  foundational:
-    [
-    {base_radius: 2, initial_difference: 0.586, converge_from_above: true, text_color: 'BLACK', feedback_background_color: 'WHITE', background_color: 'WHITE', outline_color: 'BLACK', fill_color: 'BLUE'},
-
-    {base_radius: 3, initial_difference: 0.879, converge_from_above: true, text_color: 'BLACK', feedback_background_color: 'WHITE', background_color: 'WHITE', outline_color: 'BLACK', fill_color: 'BLUE'},
-
-    {base_radius: 4, initial_difference: 1.172, converge_from_above: true, text_color: 'BLACK', feedback_background_color: 'WHITE', background_color: 'WHITE', outline_color: 'BLACK', fill_color: 'BLUE'},
-
-    {base_radius: 5, initial_difference: 1.464, converge_from_above: true, text_color: 'BLACK', feedback_background_color: 'WHITE', background_color: 'WHITE', outline_color: 'BLACK', fill_color: 'BLUE'},
-
-    {base_radius: 6, initial_difference: 1.757, converge_from_above: true, text_color: 'BLACK', feedback_background_color: 'WHITE', background_color: 'WHITE', outline_color: 'BLACK', fill_color: 'BLUE'}
-    ] 
 }
 
 const JND_CONDITIONS = {
@@ -272,226 +237,4 @@ const JND_CONDITIONS = {
     {point_color: "BLACK", strip_width: 1}
     ]  
 
-}
-
-// =========================================================
-// STEVENS DATA
-
-const STEVENS_BASE = {
-
-  foundational : 
-    [
-    {distribution_type: "gaussian", round_type: 'test', trials_per_round: 4, high_ref: 1, low_ref: 0, error: 0.0001, num_points: 100, regen_rate: 1000, mean: 0.5, SD: 0.2, num_SD: 2.5, point_color: 'BLACK', axis_color: 'BLACK', text_color: 'BLACK', background_color: 'WHITE', point_size: 3, regen_rate: 1000},
-
-    {distribution_type: "gaussian", round_type: 'test', trials_per_round: 4, high_ref: 0.5, low_ref: 0, error: 0.0001, num_points: 100, regen_rate: 1000, mean: 0.5, SD: 0.2, num_SD: 2.5, point_color: 'BLACK', axis_color: 'BLACK', text_color: 'BLACK', background_color: 'WHITE', point_size: 3, regen_rate: 1000},
-
-    {distribution_type: "gaussian", round_type: 'test', trials_per_round: 4, high_ref: 1, low_ref: 0.5, error: 0.0001, num_points: 100, regen_rate: 1000, mean: 0.5, SD: 0.2, num_SD: 2.5, point_color: 'BLACK', axis_color: 'BLACK', text_color: 'BLACK', background_color: 'WHITE', point_size: 3, regen_rate: 1000},
-
-    {distribution_type: "gaussian", round_type: 'test', trials_per_round: 4, high_ref: 0.25, low_ref: 0, error: 0.0001, num_points: 100, regen_rate: 1000, mean: 0.5, SD: 0.2, num_SD: 2.5, point_color: 'BLACK', axis_color: 'BLACK', text_color: 'BLACK', background_color: 'WHITE', point_size: 3, regen_rate: 1000},
-
-    {distribution_type: "gaussian", round_type: 'test', trials_per_round: 4, high_ref: 0.5, low_ref: 0.25, error: 0.0001, num_points: 100, regen_rate: 1000, mean: 0.5, SD: 0.2, num_SD: 2.5, point_color: 'BLACK', axis_color: 'BLACK', text_color: 'BLACK', background_color: 'WHITE', point_size: 3, regen_rate: 1000},
-
-    {distribution_type: "gaussian", round_type: 'test', trials_per_round: 4, high_ref: 0.75, low_ref: 0.5, error: 0.0001, num_points: 100, regen_rate: 1000, mean: 0.5, SD: 0.2, num_SD: 2.5, point_color: 'BLACK', axis_color: 'BLACK', text_color: 'BLACK', background_color: 'WHITE', point_size: 3, regen_rate: 1000},
-
-    {distribution_type: "gaussian", round_type: 'test', trials_per_round: 4, high_ref: 1, low_ref: 0.75, error: 0.0001, num_points: 100, regen_rate: 1000, mean: 0.5, SD: 0.2, num_SD: 2.5, point_color: 'BLACK', axis_color: 'BLACK', text_color: 'BLACK', background_color: 'WHITE', point_size: 3, regen_rate: 1000},
-    
-    {distribution_type: "gaussian", round_type: 'consistency', trials_per_round: 4, high_ref: 0.375, low_ref: 0.125, error: 0.0001, num_points: 100, regen_rate: 1000, mean: 0.5, SD: 0.2, num_SD: 2.5, point_color: 'BLACK', axis_color: 'BLACK', text_color: 'BLACK', background_color: 'WHITE', point_size: 3, regen_rate: 1000},
-    
-    {distribution_type: "gaussian", round_type: 'consistency', trials_per_round: 4, high_ref: 0.5, low_ref: 0.25, error: 0.0001, num_points: 100, regen_rate: 1000, mean: 0.5, SD: 0.2, num_SD: 2.5, point_color: 'BLACK', axis_color: 'BLACK', text_color: 'BLACK', background_color: 'WHITE', point_size: 3, regen_rate: 1000},
-    
-    {distribution_type: "gaussian", round_type: 'consistency', trials_per_round: 4, high_ref: 0.625, low_ref: 0.375, error: 0.0001, num_points: 100, regen_rate: 1000, mean: 0.5, SD: 0.2, num_SD: 2.5, point_color: 'BLACK', axis_color: 'BLACK', text_color: 'BLACK', background_color: 'WHITE', point_size: 3, regen_rate: 1000},
-    
-    {distribution_type: "gaussian", round_type: 'consistency', trials_per_round: 4, high_ref: 0.75, low_ref: 0.5, error: 0.0001, num_points: 100, regen_rate: 1000, mean: 0.5, SD: 0.2, num_SD: 2.5, point_color: 'BLACK', axis_color: 'BLACK', text_color: 'BLACK', background_color: 'WHITE', point_size: 3, regen_rate: 1000},
-    
-    {distribution_type: "gaussian", round_type: 'consistency', trials_per_round: 4, high_ref: 0.875, low_ref: 0.625, error: 0.0001, num_points: 100, regen_rate: 1000, mean: 0.5, SD: 0.2, num_SD: 2.5, point_color: 'BLACK', axis_color: 'BLACK', text_color: 'BLACK', background_color: 'WHITE', point_size: 3, regen_rate: 1000},
-    
-    {distribution_type: "gaussian", round_type: 'consistency', trials_per_round: 4, high_ref: 0.875, low_ref: 0.125, error: 0.0001, num_points: 100, regen_rate: 1000, mean: 0.5, SD: 0.2, num_SD: 2.5, point_color: 'BLACK', axis_color: 'BLACK', text_color: 'BLACK', background_color: 'WHITE', point_size: 3, regen_rate: 1000},
-    
-    {distribution_type: "gaussian", round_type: 'consistency', trials_per_round: 4, high_ref: 0.75, low_ref: 0.25, error: 0.0001, num_points: 100, regen_rate: 1000, mean: 0.5, SD: 0.2, num_SD: 2.5, point_color: 'BLACK', axis_color: 'BLACK', text_color: 'BLACK', background_color: 'WHITE', point_size: 3, regen_rate: 1000},
-    
-    {distribution_type: "gaussian", round_type: 'consistency', trials_per_round: 4, high_ref: 0.625, low_ref: 0.375, error: 0.0001, num_points: 100, regen_rate: 1000, mean: 0.5, SD: 0.2, num_SD: 2.5, point_color: 'BLACK', axis_color: 'BLACK', text_color: 'BLACK', background_color: 'WHITE', point_size: 3, regen_rate: 1000},
-    ],
-
-  design: 
-    [
-    {distribution_type: "gaussian", round_type: 'design', trials_per_round: 4, high_ref: 1, low_ref: 0, error: 0.0001, num_points: 100, regen_rate: 1000, mean: 0.5, SD: 0.2, num_SD: 2.5, point_color: 'BLACK', axis_color: 'BLACK', text_color: 'BLACK', background_color: 'WHITE', point_size: 5, regen_rate: 1000},
-
-    {distribution_type: "gaussian", round_type: 'design', trials_per_round: 4, high_ref: 1, low_ref: 0, error: 0.0001, num_points: 100, regen_rate: 1000, mean: 0.5, SD: 0.2, num_SD: 2.5, point_color: 'BLACK', axis_color: 'BLACK', text_color: 'BLACK', background_color: 'WHITE', point_size: 5, regen_rate: 1000},
-
-    {distribution_type: "gaussian", round_type: 'design', trials_per_round: 4, high_ref: 1, low_ref: 0, error: 0.0001, num_points: 100, regen_rate: 1000, mean: 0.5, SD: 0.2, num_SD: 2.5, point_color: 'BLACK', axis_color: 'BLACK', text_color: 'BLACK', background_color: 'WHITE', point_size: 5, regen_rate: 1000},
-
-    {distribution_type: "gaussian", round_type: 'design', trials_per_round: 4, high_ref: 1, low_ref: 0, error: 0.0001, num_points: 100, regen_rate: 1000, mean: 0.5, SD: 0.2, num_SD: 2.5, point_color: 'BLACK', axis_color: 'BLACK', text_color: 'BLACK', background_color: 'WHITE', point_size: 5, regen_rate: 1000},
-
-    {distribution_type: "gaussian", round_type: 'design', trials_per_round: 4, high_ref: 1, low_ref: 0, error: 0.0001, num_points: 100, regen_rate: 1000, mean: 0.5, SD: 0.2, num_SD: 2.5, point_color: 'BLACK', axis_color: 'BLACK', text_color: 'BLACK', background_color: 'WHITE', point_size: 5, regen_rate: 1000}
-    ]
-}
-
-const STEVENS_CONDITIONS = {
-
-  strip_ring_size:
-    [
-    {ring_thickness: 1},
-
-    {ring_thickness: 1},
-
-    {ring_thickness: 1},
-
-    {ring_thickness: 1},
-
-    {ring_thickness: 1},
-
-    {ring_thickness: 1},
-
-    {ring_thickness: 1},
-
-    {ring_thickness: 1},
-
-    {ring_thickness: 1},
-
-    {ring_thickness: 1},
-
-    {ring_thickness: 1},
-
-    {ring_thickness: 1},
-
-    {ring_thickness: 1},
-
-    {ring_thickness: 1},
-
-    {ring_thickness: 1}
-    ],
-
-  line_length_strip:
-  [
-    {point_color: "BLACK", strip_width: 1},
-
-    {point_color: "BLACK", strip_width: 1},
-
-    {point_color: "BLACK", strip_width: 1},
-
-    {point_color: "BLACK", strip_width: 1},
-
-    {point_color: "BLACK", strip_width: 1},
-
-    {point_color: "BLACK", strip_width: 1},
-
-    {point_color: "BLACK", strip_width: 1},
-
-    {point_color: "BLACK", strip_width: 1},
-
-    {point_color: "BLACK", strip_width: 1},
-
-    {point_color: "BLACK", strip_width: 1},
-
-    {point_color: "BLACK", strip_width: 1},
-
-    {point_color: "BLACK", strip_width: 1},
-
-    {point_color: "BLACK", strip_width: 1},
-
-    {point_color: "BLACK", strip_width: 1},
-
-    {point_color: "BLACK", strip_width: 1}
-    ]  
-}
-
-/**
- * Retrieves the data for the corresponding experiment object.
- *
- * @param  experiment  {object}       Model object of the experiment 
- *
- * @return dataset     [{assoc}, {assoc}, .... ]         
- */
-function get_data(experiment){
-
-  var dataset;
-
-  var range = experiment.range;
-  var condition = experiment.condition_name;
-  var experiment_name = experiment.constructor.name;
-
-  if (experiment_name === "JND"){
-
-    dataset = JND_BASE[range];
-
-    if (condition !== "base") {
-
-      if (!JND_CONDITIONS[condition]) {
-        throw new Error(condition + " not supported.");
-      }
-
-      dataset = create_condition_dataset(dataset, JND_CONDITIONS[condition]);
-    }
-  }
-  else if (experiment_name === "JND_Radius"){
-
-    dataset = JND_RADIUS[range];
-
-  }
-  else if (experiment_name === "Stevens"){
-
-    dataset = STEVENS_BASE[range];
-
-    if (condition !== "base") {
-
-      if (!STEVENS_CONDITIONS[condition]) {
-        throw new Error(condition + " not supported.");
-      }
-
-      dataset = create_condition_dataset(dataset, STEVENS_CONDITIONS[condition]);
-    }
-  }
-  else {
-    throw new Error(experiment + " not supported.");
-  }
-
-  // Repeat subconditions if specified
-  let result = [];
-  let repeats = SUBCONDITION_REPEATS[range][condition];
-
-  if (repeats > 0) {
-      for (let i = 0; i < repeats; i++) {
-        result = result.concat(dataset);
-      }
-  } else {
-    result = dataset;
-  }    
-
-  return result;
-}
-
-/**
- * Retrieves a smaller dataset (4 subconditions) given experiment, range and condition.
- *
- * @param  experiment  {string}            "jnd" or "stevens"   
- *         range       {string}            "foundational" or "design"         
- *         condition   {string}            Name of condition
- *
- * @return dataset     [{assoc}, {assoc}, .... ]         
- */
-function get_data_subset(experiment, range, condition) {
-
-  var dataset = get_data(experiment, range, condition);
-
-  return dataset.slice(0, 4);
-}
-
-/**
- * Appends condition-specific data to the dataset.
- *
- * @param  dataset           [{assoc}, {assoc}, .... ]     dataset with base experiment constants   
- *         condition_data    [{assoc}, {assoc}, .... ]     condition set for that experiment         
- *
- * @return dataset           [{assoc}, {assoc}, .... ]  
- **/
-function create_condition_dataset(dataset, condition_data){
-
-  var condition_dataset = [];
-
-  for (let i in dataset) {
-    let obj = Object.assign({}, dataset[i], condition_data[i]);
-    condition_dataset.push(obj);
-  }
-
-  return condition_dataset;
 }
