+++ public/scripts/experiment-properties/data/constants/jnd_radius_data.js
@@ -1,0 +1,17 @@
+export {JND_RADIUS_BASE};
+
+const JND_RADIUS_BASE = {
+
+  foundational:
+    [
+    {base_radius: 2, initial_difference: 0.586, converge_from_above: true, text_color: 'BLACK', feedback_background_color: 'WHITE', background_color: 'WHITE', outline_color: 'BLACK', fill_color: 'BLUE'},
+
+    {base_radius: 3, initial_difference: 0.879, converge_from_above: true, text_color: 'BLACK', feedback_background_color: 'WHITE', background_color: 'WHITE', outline_color: 'BLACK', fill_color: 'BLUE'},
+
+    {base_radius: 4, initial_difference: 1.172, converge_from_above: true, text_color: 'BLACK', feedback_background_color: 'WHITE', background_color: 'WHITE', outline_color: 'BLACK', fill_color: 'BLUE'},
+
+    {base_radius: 5, initial_difference: 1.464, converge_from_above: true, text_color: 'BLACK', feedback_background_color: 'WHITE', background_color: 'WHITE', outline_color: 'BLACK', fill_color: 'BLUE'},
+
+    {base_radius: 6, initial_difference: 1.757, converge_from_above: true, text_color: 'BLACK', feedback_background_color: 'WHITE', background_color: 'WHITE', outline_color: 'BLACK', fill_color: 'BLUE'}
+    ] 
+}
