+++ public/scripts/experiment-properties/distribution/gaussian_distribution_generator.js
