+++ public/scripts/experiments/stevens/stevens.js
@@ -1,8 +1,8 @@
 // import {generateDistribution} from "/scripts/generators/gaussian_distribution_generator.js";
-import {initialize_latin_square} from "/scripts/generators/latin_square_generator.js";
-import {initialize_random_order} from "/scripts/generators/random_generator.js";
+import {initialize_latin_square} from "/scripts/experiment-properties/balancing/latin_square_generator.js";
+import {initialize_random_order} from "/scripts/experiment-properties/balancing/random_generator.js";
 import {get_data, 
-        get_data_subset} from "/scripts/generators/data_generator.js";
+        get_data_subset} from "/scripts/experiment-properties/data/data_controller.js";
 import {prepare_coordinates,
         randomize_position,
         randomize_radius_position,
