+++ public/views/header.ejs
@@ -28,7 +28,7 @@
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
 
     <!-- Local scripts: -->
-    <script type="text/javascript" src="/scripts/generators/gaussian_distribution_generator.js" ></script> 
+    <script type="text/javascript" src="/scripts/experiment-properties/distribution/gaussian_distribution_generator.js" ></script> 
   </head>
 
   <body>
